LayoutTests/inspector/console/console-tests-expected.txt

CONSOLE MESSAGE: line 9: log
CONSOLE MESSAGE: line 10: debug
CONSOLE MESSAGE: line 11: info
CONSOLE MESSAGE: line 12: warn
CONSOLE MESSAGE: line 13: error
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 20: 1
CONSOLE MESSAGE: line 22: [object Object]
CONSOLE MESSAGE: line 23: 42
CONSOLE MESSAGE: line 24: true
CONSOLE MESSAGE: line 25: null
CONSOLE MESSAGE: line 26: undefined
CONSOLE MESSAGE: line 27: [object HTMLDocument]
CONSOLE MESSAGE: line 28: function () { }
CONSOLE MESSAGE: line 29: function f() { }
CONSOLE MESSAGE: line 30: 1,2,3
CONSOLE MESSAGE: line 31: /regexp.*/
Tests that console logging dumps proper messages.

console-tests.html:9log console-message console-log-level
console-tests.html:10debug console-message console-log-level
console-tests.html:11info console-message console-log-level
console-tests.html:12warn console-message console-warning-level
console-tests.html:13error console-message console-error-level
5console-tests.html:15repeated console-message console-log-level repeated-message
console-tests.html:17count: 1 console-message console-log-level
console-tests.html:17count: 2 console-message console-log-level
console-tests.html:18group console-message console-log-level console-group-title
 console-message console-log-level
console-tests.html:201 2 3 console-message console-log-level
console-tests.html:21groupCollapsed console-message console-log-level console-group-title
console-tests.html:22Object console-message console-log-level
console-tests.html:2342 console-message console-log-level
console-tests.html:24true console-message console-log-level
console-tests.html:25null console-message console-log-level
console-tests.html:26undefined console-message console-log-level
console-tests.html:27#document console-message console-log-level
console-tests.html:28function () { } console-message console-log-level
console-tests.html:29function f() { } console-message console-log-level
console-tests.html:30[1, 2, 3] console-message console-log-level
console-tests.html:31/regexp.*/ console-message console-log-level
